+++ source/extensions/omni.flux.stage_manager.plugin.listener.usd/omni/flux/stage_manager/plugin/listener/usd/base/usd_base.py
@@ -20,13 +20,17 @@ from typing import Generic, TypeVar
 
 from omni.flux.stage_manager.factory import StageManagerDataTypes as _StageManagerDataTypes
 from omni.flux.stage_manager.factory.plugins import StageManagerListenerPlugin as _StageManagerListenerPlugin
-from pydantic import Field
+from pydantic import PrivateAttr
 
 T = TypeVar("T")
 
 
 class StageManagerUSDListenerPlugin(_StageManagerListenerPlugin[T], Generic[T], abc.ABC):
-    context_name: str = Field("", description="The context name provided by the context plugin", exclude=True)
+    _context_name: str = PrivateAttr("")
+
+    def set_context_name(self, name: str):
+        """Set usd context to initialize plugin before items are rebuilt."""
+        self._context_name = name
 
     @classmethod
     @property