+++ source/extensions/omni.flux.stage_manager.plugin.widget.usd/omni/flux/stage_manager/plugin/widget/usd/action_is_visible.py
@@ -71,7 +71,7 @@ class IsVisibleActionWidgetPlugin(_StageManagerStateWidgetPlugin):
 
         self._item_clicked(button, True, model, item)
 
-        context = omni.usd.get_context(self.context_name)
+        context = omni.usd.get_context(self._context_name)
         omni.kit.commands.execute(
             "ToggleVisibilitySelectedPrims", selected_paths=context.get_selection().get_selected_prim_paths()
         )